+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Add feature flag for magic bitboard move generation.
+
 ## 2.0.0
 - Implement `Hash` for many types
 - Add new `Piece` type to make API easier to use